requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
google-generativeai==0.3.2
PyGithub==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0